refactor taskchampion::task into submodules
